requirements.txt
Pillow==6.2.1
Faker==2.0.3
requests==2.22.0
numpy==1.17.3
tensorflow==2.0.0
h5py==2.10.0
opencv-python==4.1.1.26
python-Levenshtein==0.12.0
